A kind of working build
This is not the way to do things, I'm pretty sure of that.

None the less I have a VM that boots which I am unable to
log into. The important part, a nixos VM that boots.

A key compenent, run:

$ nixos-generate-config --root /mnt

What this did was generate files specific to the VM I was
working on. I then copied those settings to the configurations
in config-nixos. They are too specific, but I'll take this
milestone and make it better.