Getting closer
There are a few things that will fail when running this
setup, but generally this is getting a lot closer to what
I believe should be happening.

Bootstrapping still needs help, but at the very least I've
learned to let nixos set the hardware settings and focus
on machine and up layers.