Mount shared vmware folder
Manually, I symlink the prj folder on the mac
to the linux vm in /home/jason. The idea is that
I can use MacOS backup software to save the projects
as I work on them.

I'm still on the fence if working on a shared folder
in the VM is what I want. So for now I'm leaving
this as a manual step.