A new re-build - a messy state
Lot going on in here that I have not yet taken
time to understand; however, in this state the
build completes.

Of note, the homemanager configuration is
disabled because it currently fails.

This version borrows more from mitchellh/nixos-config
even in cases where I know I don't want
a package in order to get the build to
work. It appears I went my own path on
the previous build. This time I wanted
to have a working build quickly which
meant I needed to start with someone
elses build.